Update umd build
