+++ examples/browser-standalone/index.html
@@ -17,8 +17,6 @@
 <script src="../../dist/tungsten.template.js"></script>
 <!-- Ractive.js for parsing templates -->
 <script src="http://cdn.ractivejs.org/latest/ractive.js"></script>
-<!-- Template helper adds tungsten.template.compileTemplates -->
-<script src="template_helper.js"></script>
 <!-- Application code -->
 <script>
     Ractive.DEBUG = false;
