+++ webpack.standalone.js
@@ -7,7 +7,7 @@ module.exports = webpackSettings({
     'backbone': path.join(__dirname, './adaptors/backbone'),
     'ampersand': path.join(__dirname, './adaptors/ampersand'),
     'core': path.join(__dirname, './tungsten'),
-    'template': [path.join(__dirname, './src/template/template.js')]
+    'template': [path.join(__dirname, './precompile/tungsten_template/template_helper.js')]
   },
   output: {
     filename: path.join(__dirname, './dist/tungsten.[name].js'),
